Use 127.0.0.1 instead of localhost for sessions.
Localhost is often configured with entries for both ipv4 (127.0.0.1) and
ipv6 entries (::1). Which one clients resolve to is inconsistent, so I
found that sometimes nodejs would end up trying to connect to the ipv6
endpoint when the session was being listened to on ipv6.

Now we just consistently listen on 127.0.0.1. Technically it's possible
for hosts to be ipv6 only or for a different loopback interface to be
pointed to by /etc/hosts (e.g. 127.0.1.1), but these cases are extremely
obscure to my knowledge so I think they can be dealt with later if
needed.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>