Refactor Caching tests to use Dagger-in-Dagger
I'm really not a fan of what I did in `core/integration/internal/engine.go`

It feels ugly and wrong and I don't like the duplication of `internal/mage/util/engine.go`. That said, I just needed it to work...

I want to/will fix it up before we merge.

Signed-off-by: Joel Longtine <joel@dagger.io>