clarify bksession args
we don't use the session name to pass along a value anymore, so we can
bring it back to identity.NewID.

the same hasn't been done for sharedKey as it's intentional semantics to
use the RouterID here so that parent/child sessions all use the same
session. though I'm not sure if it ever actually comes into play.

Signed-off-by: Alex Suraci <alex@dagger.io>