engine: skip network calls for locally cached images w/ digests
Before this, if an image reference had a digest, we'd always go
check the registry first to get metadata even if the image was already
pulled locally.

This makes no sense since image's with a digest are immutable.

Now images with a digest will be checked for in the local cache first,
with the network queried only if they aren't found there.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>