Uncached module initialization time improvements (#8761)
* go: speed up go module builds by pre-building common deps

Uncached module initializations that have lots of Go modules in the
dependency DAG are primarily bottlenecked by lots of concurrent go
operations running in parallel, which overload the CPU pretty easily.

De-duping concurrent operations is proving to be extremely tricky, but
this relatively simple change helps a bit in the meantime.

Now when we build the engine image and include the Go SDK base image,
that base image comes with the standard lib and the fixed Go SDK deps
already pulled and compiled into go's caches.

This reduces the amount of overlapping, duplicated work concurrent go
builds do when uncached.

In our own CI module, this reduces uncached initialization from ~1m25s
to ~57s, so 28s saved. This is due to CI module having lots of Go
module deps.

In the case of initializing a single Go module with no deps the
improvement is smaller; e.g. initializing `modules/alpine` goes from
~25s to ~23s.

The size increase in the engine image is significant: 165MB before vs
264MB after.

However, considering that these improvements mostly impact the
"ephemeral engine" case, which is essentially always running in CI
system with decent bandwidth, some rough calculations show that the time
to pull the engine layers will increase by 2-3s, at least in GHA. The
end result is that "lots of go sdk modules" cases will still benefit
greatly from this and "single dep-less go sdk module" will either be
about the same or possibly slightly improved.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* engine: compress SDK images w/ zstd to reduce image size + init time

Previously the SDK images included with the engine image were left
uncompressed. The idea was that the "outer" compression of the engine
image would handle compression and thus no extra layer was needed on the
cached SDK images.

However, it turns out that compressing the SDK images w/ Zstd reduces
the final engine image size and, more surprisingly, reduces uncached
module initialization time.

The engine image size is likely reduced just because the "outer"
compression is gzip but zstd is just better (simultaneously faster and
better compressed).
* `engine.tar` goes from 197MB to 191MB
* It would be nice to compress the whole engine image w/ zstd, but we've
  had problems with zstd compatibility for users on old versions of
  docker, so not pursuing that right now.

I was more surprised by the fact that the time to initialize modules in
the completely-uncached case went down too.
* Initializing our full CI module went from 1m13s to 1m7s
* Initializing `modules/alpine` by itself went from 23s to 21s

I haven't added enough spans to pinpoint why this is precisely (will do
in various follow-ups) but I'd guess that loading the SDK image from the
engine's filesystem into the content-store has enough overhead that the
smaller image actually helps even though we need to unpack it.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>