engine: fix leaked tmpdirs when subpaths of mounts used
Old code we picked up from buildkit had a bug where if a temporary mount
was created and then a subpath was the target of the mount used, the
root tempdir would never get removed. If you run a bunch of tests in an
engine and then exec is and ls tmp you will see hundreds/thousands of
empty dirs. This fixes the removal logic to always remove the actual
root of the temp dir.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>